last commit

+++ src/controllers/taskFlowController.ts
@@ -2,7 +2,8 @@ import { Request, Response, NextFunction } from "express";
 import { DefaultController } from "@AppControllers/defaultController";
 import { BadRequestError } from "@AppErrors/bad-request-error";
 import { CommonService } from "@AppServices/db/commonService";
-
+import bcrypt from 'bcrypt';
+import jwt from 'jsonwebtoken'
 /**
  * Controller class for handling user sign-up.
  * @extends DefaultController
@@ -20,21 +21,84 @@ export class taskflowController extends DefaultController {
         this._commonService = new CommonService();
 	}
 
+    async register(){
+        try{
+            const { name, email, password } = this.getRequestObj().body
+            const saltRounds = 10;
+
+
+            let hashedPassword= await bcrypt.hash(password,saltRounds)
+
+            console.log(name, email, password, hashedPassword)
+            const userModel = this.getAppObj().models.user
+            const user = await this._globalService.insertRecord(userModel,{name,email,password: hashedPassword}) as any
+
+            this.setResponseData({
+                user: {
+                    id: user.dataValues.id,
+                    name: user.dataValues.name,
+                    email: user.dataValues.email,
+                    password: user.dataValues.password
+                }
+            })
+            this.setMessage("User created successfully.");
+            this.setSuccess(true);
+            this.sendResponse();
+        }catch(e){
+            throw e
+        }
+
+    }
+
+    async login(){
+        try{
+            const { username, password } = this.getRequestObj().body
+            const userModel = this.getAppObj().models.user
+
+            let user = await this._globalService.findOne(userModel,{email:username}) as any
+
+            if(!user){
+                throw new BadRequestError("Invalid username")
+            }
+
+            let pass = await bcrypt.compare(password,user?.dataValues.password)
+
+            if(!pass){
+                throw new BadRequestError("Incorrect Password")
+            }
+
+            let token = jwt.sign({id: user.dataValues.id, name: user.dataValues.name, email: user.dataValues.email},"jgdgtfrtkdjahmnhoj")
+
+            console.log("token",token)
+
+            this.setResponseData({
+                token
+            })
+            this.setMessage("User created successfully.");
+            this.setSuccess(true);
+            this.sendResponse();
+        }catch(e){
+            throw e
+        }
+
+    }
+
     async addTask(){
         try{
             const { title, description } = this.getRequestObj().body
 
-            const taskModel = this.getAppObj().models.taskflow
-            const task = await this._globalService.insertRecord(taskModel,{title,description}) as any
+            const user = this.getRequestObj().body.user
 
-            console.log(task.dataValues)
+            const taskModel = this.getAppObj().models.taskflow
+            const task = await this._globalService.insertRecord(taskModel,{title,description,user_id: user.id}) as any
 
             this.setResponseData({
                 task: {
                     id: task.dataValues.id,
                     title: task.dataValues.title,
                     description: task.dataValues.description,
-                    status: task.dataValues.status
+                    status: task.dataValues.status,
+                    createdBy: user.name
                 }
             })
             this.setMessage("Task created successfully.");
@@ -46,10 +110,22 @@ export class taskflowController extends DefaultController {
 
     }
 
+
     async getAllTask(){
         try{
             const taskModel = this.getAppObj().models.taskflow
-            const tasks = await this._globalService.findAllRecords(taskModel,{},undefined,[],["created_at"]) as any
+            const status = this.getRequestObj().query.status
+            let user = this.getRequestObj().body.user
+
+            let conditions: any = {
+                user_id: user.id.toString()
+            }
+
+            if(status){
+                conditions["status"] = status
+            }
+
+            const tasks = await this._globalService.findAllRecords(taskModel,conditions,undefined,[],["created_at"]) as any
 
             this.setResponseData({
                 tasks
@@ -66,6 +142,7 @@ export class taskflowController extends DefaultController {
     async updateTask(){
         try{
             const id = this.getRequestObj().params.id
+            let user = this.getRequestObj().body.user
             if(!id){
                 throw new BadRequestError("Id is required")
             }
@@ -74,7 +151,7 @@ export class taskflowController extends DefaultController {
             let task
 
             const taskModel = this.getAppObj().models.taskflow
-            const updateCount = await this._globalService.updateRecords(taskModel,{status},{id})
+            const updateCount = await this._globalService.updateRecords(taskModel,{status},{id, user_id: user.id})
             if(updateCount == 0){
                 throw new BadRequestError("Task not found")
             }else{
@@ -104,9 +181,10 @@ export class taskflowController extends DefaultController {
             if(!id){
                 throw new BadRequestError("Id is required")
             }
+            let user = this.getRequestObj().body.user
 
             const taskModel = this.getAppObj().models.taskflow
-            const deleteCount = await this._globalService.deleteRecords(taskModel,{id})
+            const deleteCount = await this._globalService.deleteRecords(taskModel,{id, user_id: user.id})
             if(deleteCount["row-deleted"] == 0){
                 throw new BadRequestError("Task not found")
             }
